MenusDialogをコンポーネントとして切り出す
