+++ src/common/layout/index.tsx
@@ -1,10 +1,14 @@
 import React from 'react';
-import Header from '../components/molecules/Header';
+import { Header, HeaderProps } from '../components/molecules/Header';
 import styled from 'styled-components';
 
 const Wrapper = styled.div``;
 
-const MyLayout = (props: { children: React.ReactNode; pathname: string }) => (
+type MyLayoutProps = HeaderProps & {
+  children: React.ReactNode;
+};
+
+const MyLayout = (props: MyLayoutProps) => (
   <Wrapper style={{ flex: 1 }}>
     <Header pathname={props.pathname} />
     {props.children}
